feat: typescript and tailwind
