--- components/under-construction/styles.module.css
@@ -1,82 +1,0 @@
-.home {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-family: Helvetica, Arial, sans-serif;
-  height: 100dvh;
-}
-
-.container {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  gap: 78px;
-}
-
-.titleContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 8px;
-}
-
-.home h1 {
-  font-size: 40px;
-  font-weight: 900;
-  padding: 0;
-  margin: 0;
-}
-
-.home p {
-  padding: 4px;
-}
-
-.buttonLink {
-  font-size: larger;
-  font-weight: 700;
-  text-decoration: none;
-  background-color: #f0da5e;
-  color: #383838;
-  padding: 10px;
-  border-radius: 8px;
-}
-
-.buttonLink:hover {
-  opacity: 0.8;
-}
-
-.textContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 4px;
-}
-
-.textContainer p {
-  margin: 0;
-}
-
-.warningWrapper {
-  display: flex;
-  flex-direction: column;
-  border-radius: 8px;
-  padding: 12px 18px;
-  background-color: gainsboro;
-  gap: 16px;
-}
-
-.warningWrapper h2 {
-  font-size: 16px;
-}
-
-.warningWrapper span {
-  font-size: 14px;
-}
-
-.endContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 8px;
-}
