--- infra/database.js
@@ -1,51 +1,0 @@
-import { Client } from "pg";
-
-function getSSLValues() {
-  if (process.env.POSTGRES_CA) {
-    return {
-      ca: process.env.POSTGRES_CA,
-    };
-  }
-
-  return process.env.NODE_ENV === "production";
-}
-
-async function getNewClient() {
-  const client = new Client({
-    host: process.env.POSTGRES_HOST,
-    port: process.env.POSTGRES_PORT,
-    user: process.env.POSTGRES_USER,
-    database: process.env.POSTGRES_DB,
-    password: process.env.POSTGRES_PASSWORD,
-    ssl: getSSLValues(),
-  });
-
-  await client.connect();
-
-  return client;
-}
-
-async function query(queryObject) {
-  let client;
-
-  try {
-    client = await getNewClient();
-
-    const response = await client.query(queryObject);
-
-    return response;
-  } catch (error) {
-    console.error(error);
-
-    throw error;
-  } finally {
-    await client.end();
-  }
-}
-
-const database = {
-  query,
-  getNewClient,
-};
-
-export default database;
