+++ jest.config.js
@@ -5,11 +5,11 @@ dotenv.config({ path: ".env.development" });
 const nextJest = require("next/jest");
 
 const createJestConfig = nextJest({
-  dir: ".",
+  dir: "src",
 });
 
 const jestConfig = createJestConfig({
-  moduleDirectories: ["node_modules", "<rootDir>"],
+  moduleDirectories: ["node_modules", "<rootDir>", "src"],
   testTimeout: 60_000,
 });
