+++ package.json
@@ -25,6 +25,9 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
+    "@tailwindcss/postcss": "4.1.5",
+    "@types/async-retry": "1.4.9",
+    "@types/react": "19.1.2",
     "@vercel/analytics": "1.1.0",
     "async-retry": "1.3.3",
     "dotenv": "16.4.5",
@@ -34,12 +37,16 @@
     "ora": "8.0.1",
     "pg": "8.12.0",
     "react": "18.3.1",
-    "react-dom": "18.3.1"
+    "react-dom": "18.3.1",
+    "react-icons": "^5.5.0"
   },
   "devDependencies": {
     "@commitlint/cli": "19.4.0",
     "@commitlint/config-conventional": "19.2.2",
+    "@types/css-modules": "1.0.5",
     "@types/jest": "29.5.12",
+    "@types/node": "18.19.87",
+    "autoprefixer": "10.4.21",
     "commitizen": "4.3.1",
     "concurrently": "8.2.2",
     "cz-conventional-changelog": "3.3.0",
@@ -49,7 +56,11 @@
     "eslint-plugin-jest": "28.8.0",
     "husky": "9.1.4",
     "jest": "29.7.0",
-    "prettier": "3.3.3"
+    "postcss": "8.5.3",
+    "prettier": "^3.3.3",
+    "prettier-plugin-tailwindcss": "0.6.11",
+    "tailwindcss": "3.4.17",
+    "typescript": "5.8.3"
   },
   "config": {
     "commitizen": {
