--- pages/index.js
@@ -1,7 +1,0 @@
-import { UnderConstruction } from "../components/under-construction";
-
-function Home() {
-  return <UnderConstruction />;
-}
-
-export default Home;
