+++ src/database.ts
@@ -1,0 +1,1 @@
+export { database } from "../infra/database";
