+++ src/pages/status/index.tsx
@@ -1,0 +1,13 @@
+import { Header } from "components";
+
+export default function StatusPage() {
+  return (
+    <div className="flex h-dvh flex-col">
+      <Header />
+
+      <div className="flex h-full w-full items-center justify-center">
+        <span className="opacity-50">status page placeholder</span>
+      </div>
+    </div>
+  );
+}
