+++ src/tests/orchestrator.ts
@@ -1,0 +1,30 @@
+import { database } from "database";
+import retry from "async-retry";
+
+async function waitForAllServices() {
+  async function waitForWebServer() {
+    async function fetchStatusPage() {
+      const response = await fetch("http://localhost:3000/api/v1/status");
+
+      if (response.status !== 200) {
+        throw Error();
+      }
+    }
+
+    return retry(fetchStatusPage, {
+      retries: 100,
+      maxTimeout: 1000,
+    });
+  }
+
+  await waitForWebServer();
+}
+
+async function clearDatabase() {
+  await database.query("drop schema public cascade; create schema public;");
+}
+
+export const orchestrator = {
+  waitForAllServices,
+  clearDatabase,
+};
