--- styles/global.css
@@ -1,5 +1,0 @@
-* {
-  box-sizing: border-box;
-  padding: 0;
-  margin: 0;
-}
